% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ehr_data}
\alias{ehr_data}
\title{A Synthetic EHR Dataset}
\format{A data.frame with 10000 observations of 588 variables.}
\usage{
data(ehr_data)
}
\description{
This dataset gives a sample dataset for EHR phenotyping.
It contains counts for ICD codes, counts for NLP mentions,
healthcare utilization (HU) features for all observations.
It also contains the accurate phenotypes for 181 observations.
}
\keyword{datasets}
