% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ehrdata}
\alias{ehrdata}
\title{Example EHR data with 10,000 records.}
\format{A data frame with 10,000 rows and 588 variables:}
\usage{
ehrdata
}
\description{
A dataset containing the example to illustrate the main steps of the 
algorithms and usage of functions in the package.
}
\keyword{datasets}
