% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{PheCAP-package}
\alias{PheCAP-package}
\title{High-Throughput Phenotyping with Electronic Health Records 
using a Common Automated Pipeline}
\description{
Implement surrogate-assisted feature extraction (SAFE) and common 
machine learning approaches to train and validate phenotyping models.
Background and details about the methods can be found at 
<doi:10.1093/jamia/ocw135> and <doi:10.1136/bmj.h1885>.
}
\details{
PheCAP provides a straightforward interface for conducting
phenotyping on eletronic health records. One can specify the
data via \code{\link{PhecapData}}, define surrogate using
\code{\link{PhecapSurrogate}}. Next, one may run
surrogate-assisted feature extraction (SAFE) by calling
\code{\link{phecap_run_feature_extraction}}, and then
train and validate phenotyping models via
\code{\link{phecap_train_phenotyping_model}} and
\code{\link{phecap_validate_phenotyping_model}}.
The predictive performance can be visualized using
\code{\link{phecap_plot_roc_curves}}.
Predicted phenotype is provided by
\code{\link{phecap_predict_phenotype}}.
}
\keyword{package}
