% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{phecap_predict_phenotype}
\alias{phecap_predict_phenotype}
\title{Predict Phenotype}
\usage{
phecap_predict_phenotype(data, model)
}
\arguments{
\item{data}{an object of class \code{PhecapData}, obtained by calling \code{PhecapData(...)}.}

\item{model}{an object of class \code{PhecapModel}, probably returned from
\code{phecap_train_phenotyping_model}.}
}
\value{
A \code{data.frame} with two columns:
\item{patient_index}{patient identifier},
\item{prediction}{predicted phenotype}.
}
\description{
Compute predicted probability of having the phenotype
for each patient in the dataset.
}
