% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PhecapData}
\alias{PhecapData}
\title{Define or Read Datasets for Phenotyping}
\usage{
PhecapData(icd_feature, nlp_feature, hu_feature, training_label = NULL,
  validation_label = NULL, patient_index = "PatientID",
  feature_transformation = log1p)
}
\arguments{
\item{icd_feature}{a character scalar of the path of ICD feature, 
or a data.frame of the actual ICD feature.}

\item{nlp_feature}{a character scalar of the path of NLP feature, 
or a data.frame of the actual NLP feature.}

\item{hu_feature}{a character scalar of the path of HU feature, 
or a data.frame of the actual HU feature. One may also include other 
variables that must be included in the model.}

\item{training_label}{a character scalar of the path of training label, 
or a data.frame of the actual training label, or NULL.}

\item{validation_label}{a character scalar of the path of validation label, 
or a data.frame of the actual validation label, or NULL.}

\item{patient_index}{a character scalar or vector that specify a unique identifier 
for each patient in the data. Such variable(s) must appear in all the dataset above.}

\item{feature_transformation}{a function that will be applied to all the features. 
Since count data are typically right-skewed, by default \code{log1p} will be used.}
}
\value{
An object of class \code{PhecapData}.
}
\description{
Specify the data to be used for phenotyping.
}
\details{
Note that icd_feature, nlp_feature and hu_feature cannot be NULL, while training_label
can be NULL if only feature extraction is needed, and validation_label can be NULL 
if only feature extraction and model training is needed.
}
