\name{ehr_data}
\docType{data}
\alias{ehr_data}
\title{A Synthetic EHR Dataset}
\description{
  This dataset gives a sample dataset for EHR phenotyping.
  It contains counts for ICD codes, counts for NLP mentions,
  healthcare utilization (HU) features for all observations.
  It also contains the accurate phenotypes for 181 observations.
}
\usage{data(ehr_data)}
\format{A data.frame with 10000 observations of 588 variables.}
\keyword{datasets}
